+++ db/migrate/10_add_super_auth_label_to_resources.rb
@@ -1,0 +1,13 @@
+Sequel.migration do
+  up do
+    alter_table(:super_auth_resources) do
+      add_column :super_auth_label, String
+    end
+  end
+
+  down do
+    alter_table(:super_auth_resources) do
+      drop_column :super_auth_label
+    end
+  end
+end