+++ db/migrate_activerecord/20250101000010_add_super_auth_label_to_super_auth_resources.rb
@@ -1,0 +1,5 @@
+class AddSuperAuthLabelToSuperAuthResources < ActiveRecord::Migration[7.0]
+  def change
+    add_column :super_auth_resources, :super_auth_label, :string
+  end
+end